refactor: rename configJsons dir to config dir
