+++ config/gametitle.json
@@ -1,0 +1,5 @@
+{
+    "VALORANT": "https://static-cdn.jtvnw.net/ttv-boxart/516575-720x960.jpg",
+    "League of Legends": "https://static-cdn.jtvnw.net/ttv-boxart/21779-720x960.jpg",
+    "Minecraft": "https://static-cdn.jtvnw.net/ttv-boxart/27471_IGDB-720x960.jpg"
+}
