--- configJsons/config.json
@@ -1,3 +1,0 @@
-{
-    "prefix": "!"
-}
